Move login form into pages directory
